style: Resolve all eslint errors
