+++ app/tests/test_helper.js
@@ -11,7 +11,7 @@ global.window = win;
 // from mocha-jsdom https://github.com/rstacruz/mocha-jsdom/blob/master/index.js#L80
 Object.keys(window).forEach((key) => {
   if (!(key in global)) {
-    global[key] = window[key];
+    global[key] = window[key];
   }
 });
